Move storeArticle to App; Drill storeArticle to TopicCard
