+++ src/components/Article/Article.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Link } from 'react-router-dom'
 import './Article.css';
