+++ src/components/Form/Form.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useHistory } from 'react-router-dom';
 import './Form.css';
