+++ src/components/History/History.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Link } from 'react-router-dom';
 import Card from '../Card/Card';
 import { CleanedArticle } from '../../Models';
@@ -6,10 +5,11 @@ import '../Feed/Feed.css'
 
 interface HistoryProps {
   history: CleanedArticle[] | any
-  updateHistory: (localHistory: string[]) => void;
+  updateUserSentiment: (userSentiment: number) => void
+  storeArticle: (id: string) => void
 }
 
-const History = ({ history, updateHistory }: HistoryProps): JSX.Element => {
+const History = ({ history, storeArticle, updateUserSentiment }: HistoryProps): JSX.Element => {
 
   const articleCards = history.map((article: CleanedArticle) => {
     return (
@@ -18,8 +18,9 @@ const History = ({ history, updateHistory }: HistoryProps): JSX.Element => {
         image={ article.multimedia.url }
         key={ article.title }
         sentiment={ article.sentiment }
-        id= { article.id }
-        updateHistory = { updateHistory }
+        id={ article.id }
+        storeArticle={ storeArticle }
+        updateUserSentiment={ updateUserSentiment }
       />
     )
   })
