+++ src/components/TopicCard/TopicCard.tsx
@@ -8,17 +8,23 @@ interface TopicCard {
   sentiment: number
   updateUserSentiment: (userSentiment: number) => void
   selectedTopic: string
-  updateHistory: (localHistory: string[]) => void;
+  storeArticle: (id: string) => void
 }
 
-const TopicCard = ({ title, image, id, sentiment, updateUserSentiment, selectedTopic }: TopicCard): JSX.Element => {
+const TopicCard = ({ title, image, id, sentiment, updateUserSentiment, selectedTopic, storeArticle }: TopicCard): JSX.Element => {
+
+  const handleClick = () => {
+    updateUserSentiment!(sentiment)
+    storeArticle(id)
+  }
+
   return (
 
       <div className="card-container">
         <article className="news-card cy-article-card">
           <Link
             to={`/feed/${selectedTopic}/${id}`}
-            onClick={() => updateUserSentiment(sentiment) }
+            onClick={ handleClick }
             className="cy-article-link"
           >
             <img className="article-image cy-article-image" src={image} alt={title} />
